Stub out `utils` package.
