+++ packages/utils/.npmignore
@@ -1,0 +1,2 @@
+.env
+tsconfig.tsbuildinfo
