+++ packages/utils/src/index.ts
@@ -1,0 +1,7 @@
+/**
+ * @license
+ * Copyright 2025 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export type * from "./typed-event-target.js";
